+++ packages/core/src/index.ts
@@ -10,5 +10,6 @@ export * from "./convert.js";
 export * from "./futures.js";
 export * from "./notifications.js";
 export * from "./referrals.js";
+export * from "./vip.js";
 export * from "./wallet/derivation.js";
 export * from "./wallet/provision.js";